requirements.txt
apache-airflow==2.8.0
psycopg2-binary==2.9.7
yfinance==0.2.28
requests==2.31.0
pandas==1.5.3
